Day 9, Part 1, more progress
I've swapped to using 64-bit integers, which was painful to do (much
stuff in Java is 32-bit integers internally, so there were surprising
places that I had to down-cast back to integers). Nothing terribly
interesting, but examples 2 and 3 now work.

+++ Day7.java
@@ -113,8 +113,8 @@ class Day7
 			ic.SetOps(aNProgram);
 			ic.RunOps();
 
-			int[] aNOutput = ic.ANOutput();
-			nCur = aNOutput[0];
+			long[] aNOutput = ic.ANOutput();
+			nCur = (int)aNOutput[0];
 		}
 
 		return nCur;
@@ -176,7 +176,7 @@ class Day7
 				ic.AddInput(nCur);
 				ic.RunOps();
 
-				nCur = ic.NOutputLast();
+				nCur = (int)ic.NOutputLast();
 			}
 		}
